pyproject.toml
[project]
name = "notebook-lm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastembed>=0.7.3",
    "ipykernel>=6.30.1",
    "piper-tts>=0.0.2",
    "pymilvus[milvus-lite]>=2.6.2",
    "pymupdf>=1.26.4",
    "python-dotenv>=1.1.1",
    "soundfile>=0.12.1",
    "streamlit>=1.28.0",
]
